Rename file_async to file
