+++ crates/extensions/c8y_firmware_manager/src/lib.rs
@@ -27,8 +27,8 @@ use tedge_api::path::DataDir;
 use tedge_config::models::TopicPrefix;
 use tedge_mqtt_ext::MqttMessage;
 use tedge_mqtt_ext::TopicFilter;
-use tedge_utils::file_async::create_directory_with_defaults;
-use tedge_utils::file_async::FileError;
+use tedge_utils::file::create_directory_with_defaults;
+use tedge_utils::file::FileError;
 use worker::IdDownloadRequest;
 use worker::IdDownloadResult;
 use worker::OperationOutcome;
