+++ crates/extensions/tedge_downloader_ext/src/actor.rs
@@ -14,7 +14,7 @@ use tedge_actors::Sequential;
 use tedge_actors::Server;
 use tedge_actors::ServerActorBuilder;
 use tedge_actors::ServerConfig;
-use tedge_utils::file_async::PermissionEntry;
+use tedge_utils::file::PermissionEntry;
 
 #[derive(Debug, Clone, Eq, PartialEq)]
 pub struct DownloadRequest {
