+++ plugins/c8y_remote_access_plugin/src/lib.rs
@@ -10,9 +10,9 @@ use std::io;
 use std::process::Stdio;
 use tedge_config::log_init;
 use tedge_config::TEdgeConfig;
-use tedge_utils::file_async::change_user_and_group;
-use tedge_utils::file_async::create_directory_with_user_group;
-use tedge_utils::file_async::create_file_with_user_group;
+use tedge_utils::file::change_user_and_group;
+use tedge_utils::file::create_directory_with_user_group;
+use tedge_utils::file::create_file_with_user_group;
 use tokio::io::AsyncBufReadExt;
 use tokio::io::AsyncWriteExt;
 use tokio::io::BufReader;
